Update directory structure and file naming (#282)
