+++ src/AzureOpenAIProxy.ApiApp/Endpoints/AdminEventEndpoints.cs
@@ -5,35 +5,36 @@ using Microsoft.AspNetCore.Mvc;
 namespace AzureOpenAIProxy.ApiApp.Endpoints;
 
 /// <summary>
-/// This represents the endpoint entity for get event details by admin
+/// This represents the endpoint entity for event details by admin
 /// </summary>
 public static class AdminEventEndpoints
 {
     /// <summary>
-    /// Adds the get event details by admin endpoint
+    /// Adds the admin event endpoint
     /// </summary>
     /// <param name="app"><see cref="WebApplication"/> instance.</param>
     /// <returns>Returns <see cref="RouteHandlerBuilder"/> instance.</returns>
-    public static RouteHandlerBuilder AddAdminEvents(this WebApplication app)
+    public static RouteHandlerBuilder AddNewAdminEvent(this WebApplication app)
     {
-        // Todo: Issue #19 https://github.com/aliencube/azure-openai-sdk-proxy/issues/19
-        // Need authorization by admin
-        var builder = app.MapGet(AdminEndpointUrls.AdminEventDetails, (
-            [FromRoute] string eventId) =>
+        var builder = app.MapPost(AdminEndpointUrls.AdminEvents, async (
+            [FromBody] AdminEventDetails payload,
+            HttpRequest request) =>
         {
-            // Todo: Issue #208 https://github.com/aliencube/azure-openai-sdk-proxy/issues/208
-            return Results.Ok();
-            // Todo: Issue #208
+            return await Task.FromResult(Results.Ok());
         })
+        // TODO: Check both request/response payloads
+        .Accepts<AdminEventDetails>(contentType: "application/json")
         .Produces<AdminEventDetails>(statusCode: StatusCodes.Status200OK, contentType: "application/json")
+        // TODO: Check both request/response payloads
+        .Produces(statusCode: StatusCodes.Status400BadRequest)
         .Produces(statusCode: StatusCodes.Status401Unauthorized)
         .Produces<string>(statusCode: StatusCodes.Status500InternalServerError, contentType: "text/plain")
         .WithTags("admin")
-        .WithName("GetAdminEventDetails")
+        .WithName("CreateAdminEvent")
         .WithOpenApi(operation =>
         {
-            operation.Summary = "Gets event details from the given event ID";
-            operation.Description = "This endpoint gets the event details from the given event ID.";
+            operation.Summary = "Create admin event";
+            operation.Description = "Create admin event";
 
             return operation;
         });
@@ -46,7 +47,7 @@ public static class AdminEventEndpoints
     /// </summary>
     /// <param name="app"><see cref="WebApplication"/> instance.</param>
     /// <returns>Returns <see cref="RouteHandlerBuilder"/> instance.</returns>
-    public static RouteHandlerBuilder AddAdminEventList(this WebApplication app)
+    public static RouteHandlerBuilder AddListAdminEvents(this WebApplication app)
     {
         // Todo: Issue #19 https://github.com/aliencube/azure-openai-sdk-proxy/issues/19
         // Need authorization by admin
@@ -73,11 +74,43 @@ public static class AdminEventEndpoints
     }
 
     /// <summary>
+    /// Adds the get event details by admin endpoint
+    /// </summary>
+    /// <param name="app"><see cref="WebApplication"/> instance.</param>
+    /// <returns>Returns <see cref="RouteHandlerBuilder"/> instance.</returns>
+    public static RouteHandlerBuilder AddGetAdminEvent(this WebApplication app)
+    {
+        // Todo: Issue #19 https://github.com/aliencube/azure-openai-sdk-proxy/issues/19
+        // Need authorization by admin
+        var builder = app.MapGet(AdminEndpointUrls.AdminEventDetails, (
+            [FromRoute] string eventId) =>
+        {
+            // Todo: Issue #208 https://github.com/aliencube/azure-openai-sdk-proxy/issues/208
+            return Results.Ok();
+            // Todo: Issue #208
+        })
+        .Produces<AdminEventDetails>(statusCode: StatusCodes.Status200OK, contentType: "application/json")
+        .Produces(statusCode: StatusCodes.Status401Unauthorized)
+        .Produces<string>(statusCode: StatusCodes.Status500InternalServerError, contentType: "text/plain")
+        .WithTags("admin")
+        .WithName("GetAdminEvent")
+        .WithOpenApi(operation =>
+        {
+            operation.Summary = "Gets event details from the given event ID";
+            operation.Description = "This endpoint gets the event details from the given event ID.";
+
+            return operation;
+        });
+
+        return builder;
+    }
+
+    /// <summary>
     /// Adds the update event details by admin endpoint
     /// </summary>
     /// <param name="app"><see cref="WebApplication"/> instance.</param>
     /// <returns>Returns <see cref="RouteHandlerBuilder"/> instance.</returns>
-    public static RouteHandlerBuilder AddUpdateAdminEvents(this WebApplication app)
+    public static RouteHandlerBuilder AddUpdateAdminEvent(this WebApplication app)
     {
         // Todo: Issue #19 https://github.com/aliencube/azure-openai-sdk-proxy/issues/19
         // Need authorization by admin
@@ -94,44 +127,11 @@ public static class AdminEventEndpoints
         .Produces(statusCode: StatusCodes.Status404NotFound)
         .Produces<string>(statusCode: StatusCodes.Status500InternalServerError, contentType: "text/plain")
         .WithTags("admin")
-        .WithName("UpdateAdminEventDetails")
+        .WithName("UpdateAdminEvent")
         .WithOpenApi(operation =>
         {
             operation.Summary = "Updates event details from the given event ID";
             operation.Description = "This endpoint updates the event details from the given event ID.";
-
-            return operation;
-        });
-
-        return builder;
-    }
-    
-    /// <summary>
-    /// Adds the admin event endpoint
-    /// </summary>
-    /// <param name="app"><see cref="WebApplication"/> instance.</param>
-    /// <returns>Returns <see cref="RouteHandlerBuilder"/> instance.</returns>
-    public static RouteHandlerBuilder CreateAdminEvent(this WebApplication app)
-    {
-        var builder = app.MapPost(AdminEndpointUrls.AdminEvents, async (
-            [FromBody] AdminEventDetails payload,
-            HttpRequest request) =>
-        {
-            return await Task.FromResult(Results.Ok());
-        })
-        // TODO: Check both request/response payloads
-        .Accepts<AdminEventDetails>(contentType: "application/json")
-        .Produces<AdminEventDetails>(statusCode: StatusCodes.Status200OK, contentType: "application/json")
-        // TODO: Check both request/response payloads
-        .Produces(statusCode: StatusCodes.Status400BadRequest)
-        .Produces(statusCode: StatusCodes.Status401Unauthorized)
-        .Produces<string>(statusCode: StatusCodes.Status500InternalServerError, contentType: "text/plain")
-        .WithTags("admin")
-        .WithName("CreateAdminEvent")
-        .WithOpenApi(operation =>
-        {
-            operation.Summary = "Create admin event";
-            operation.Description = "Create admin event";
 
             return operation;
         });
