+++ src/AzureOpenAIProxy.ApiApp/Endpoints/PlaygroundEndpointUrls.cs
@@ -1,0 +1,15 @@
+﻿namespace AzureOpenAIProxy.ApiApp.Endpoints;
+
+/// <summary>
+/// This represents the collection of the playground endpoint URLs.
+/// </summary>
+public static class PlaygroundEndpointUrls
+{
+    /// <summary>
+    /// Declares the event endpoint.
+    /// </summary>
+    /// <remarks>
+    /// - GET method for listing all events
+    /// </remarks>
+    public const string Events = "/events";
+}
